chore: move service and common files
