+++ packages/common/store/middlewares/loadingBar.test.js
@@ -1,6 +1,6 @@
 import { expect } from 'chai';
 import { spy } from 'sinon';
-import actionTypes from '../actions/actionTypes';
+import actionTypes from 'src/modules/common/store/actionTypes';
 import middleware from './loadingBar';
 
 describe('LoadingBar middleware', () => {
