+++ src/locales/en/common.json
@@ -162,6 +162,7 @@
   "Enter your password": "Enter your password",
   "Enter your secret recovery phrase to manage your account.": "Enter your secret recovery phrase to manage your account.",
   "Error": "Error",
+  "Error retrieving conversion rates.": "Error retrieving conversion rates.",
   "Expand sidebar": "Expand sidebar",
   "Explorer": "Explorer",
   "Failed to load block details.": "Failed to load block details.",
