+++ src/modules/common/store/actionTypes.js
