+++ src/modules/transaction/store/actions.js
@@ -5,7 +5,7 @@ import { extractKeyPair } from '@wallet/utils/account';
 import { getTransactionSignatureStatus } from '@wallet/components/signMultisigView/helpers';
 import { selectActiveTokenAccount } from '@common/store';
 import { timerReset } from '@auth/store/action';
-import { loadingStarted, loadingFinished } from '@common/store/actions/loading';
+import { loadingStarted, loadingFinished } from 'src/modules/common/store/actions';
 import actionTypes from './actionTypes';
 import { getTransactions, broadcast } from '../api';
 import {
